refactor: model 세부 패키지 분리
