+++ src/main/kotlin/blackjack/model/user/Hand.kt
@@ -1,4 +1,7 @@
-package blackjack.model
+package blackjack.model.user
+
+import blackjack.model.card.Card
+import blackjack.model.game.Score
 
 class Hand(cards: Set<Card> = emptySet()) {
     private val _cards: MutableSet<Card> = cards.toMutableSet()
