+++ src/test/kotlin/blackjack/CardDeckTest.kt
@@ -1,6 +1,6 @@
 package blackjack
 
-import blackjack.model.CardDeck
+import blackjack.model.card.CardDeck
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
