+++ src/test/kotlin/blackjack/ScoreTest.kt
@@ -1,9 +1,9 @@
 package blackjack
 
-import blackjack.model.Card
-import blackjack.model.CardNumber
-import blackjack.model.CardSymbol
-import blackjack.model.Score
+import blackjack.model.card.Card
+import blackjack.model.card.CardNumber
+import blackjack.model.card.CardSymbol
+import blackjack.model.game.Score
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
